+++ lib/searchkick/query.rb
@@ -459,17 +459,25 @@ module Searchkick
 
     def set_boost_by_distance(custom_filters)
       boost_by_distance = options[:boost_by_distance] || {}
-      boost_by_distance = {function: :gauss, scale: "5mi"}.merge(boost_by_distance)
-      if !boost_by_distance[:field] || !boost_by_distance[:origin]
-        raise ArgumentError, "boost_by_distance requires :field and :origin"
+
+      # legacy format
+      if boost_by_distance[:field]
+        boost_by_distance = {boost_by_distance[:field] => boost_by_distance.except(:field)}
       end
-      function_params = boost_by_distance.select { |k, _| [:origin, :scale, :offset, :decay].include?(k) }
-      function_params[:origin] = location_value(function_params[:origin])
-      custom_filters << {
-        boost_by_distance[:function] => {
-          boost_by_distance[:field] => function_params
+
+      boost_by_distance.each do |field, attributes|
+        attributes = {function: :gauss, scale: "5mi"}.merge(attributes)
+        unless attributes[:origin]
+          raise ArgumentError, "boost_by_distance requires :origin"
+        end
+        function_params = attributes.select { |k, _| [:origin, :scale, :offset, :decay].include?(k) }
+        function_params[:origin] = location_value(function_params[:origin])
+        custom_filters << {
+          attributes[:function] => {
+            field => function_params
+          }
         }
-      }
+      end
     end
 
     def set_boost_by(multiply_filters, custom_filters)