+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `retain` option to `reindex`
 - Added support for attributes in highlight tags
 - Fixed potentially silent errors in reindex job
+- Improved format for `boost_by_distance`
 
 ## 2.0.1